backend/requirements.txt
asgiref==3.11.1
cffi==2.0.0
cryptography==46.0.5
dj-database-url==3.1.2
Django==6.0.2
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
gunicorn==25.1.0
packaging==26.0
pycparser==3.0
PyJWT==2.11.0
PyMySQL==1.1.2
sqlparse==0.5.5
tzdata==2025.3
whitenoise==6.12.0
